pyproject.toml
[project]
name = "emergency-call-engine"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.6",
    "hupper>=1.12.1",
    "pika>=1.3.2",
    "pymongo>=4.10.1",
    "python-dotenv>=1.0.1",
    "sqlalchemy>=2.0.36",
]
